# docs/api_openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VRS Nesting Web API",
    "version": "0.3.1-phase3-fixes"
  },
  "paths": {
...
    "/v1/projects/{project_id}/run-configs": {
      "post": {
        "tags": [
          "run-configs"
        ],
        "summary": "Create Run Config",
        "operationId": "create_run_config_v1_projects__project_id__run_configs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunConfigCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "run-configs"
        ],
        "summary": "List Run Configs",
        "operationId": "list_run_configs_v1_projects__project_id__run_configs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunConfigListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_v1_projects__project_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_v1_projects__project_id__runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "done",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_v1_projects__project_id__runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_v1_projects__project_id__runs__run_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InlineRunConfig": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schema_version": {
            "type": "string",
            "maxLength": 40,
            "title": "Schema Version",
            "default": "dxf_v1"
          },
          "seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seed",
            "default": 0
          },
          "time_limit_s": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Time Limit S",
            "default": 60
          },
          "spacing_mm": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Spacing Mm",
            "default": 2.0
          },
          "margin_mm": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Margin Mm",
            "default": 5.0
          },
          "stock_file_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stock File Id"
          },
          "parts_config": {
            "items": {
              "$ref": "#/components/schemas/RunConfigPartEntry"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Parts Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stock_file_id",
          "parts_config"
        ],
        "title": "InlineRunConfig"
      },
...
      "RunConfigCreateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schema_version": {
            "type": "string",
            "maxLength": 40,
            "title": "Schema Version",
            "default": "dxf_v1"
          },
          "seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seed",
            "default": 0
          },
          "time_limit_s": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Time Limit S",
            "default": 60
          },
          "spacing_mm": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Spacing Mm",
            "default": 2.0
          },
          "margin_mm": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Margin Mm",
            "default": 5.0
          },
          "stock_file_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stock File Id"
          },
          "parts_config": {
            "items": {
              "$ref": "#/components/schemas/RunConfigPartEntry"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Parts Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stock_file_id",
          "parts_config"
        ],
        "title": "RunConfigCreateRequest"
      },
      "RunConfigListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunConfigResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RunConfigListResponse"
      },
      "RunConfigPartEntry": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id"
          },
          "quantity": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          },
          "allowed_rotations_deg": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Allowed Rotations Deg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "file_id"
        ],
        "title": "RunConfigPartEntry"
      },
      "RunConfigResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "time_limit_s": {
            "type": "integer",
            "title": "Time Limit S"
          },
          "spacing_mm": {
            "type": "number",
            "title": "Spacing Mm"
          },
          "margin_mm": {
            "type": "number",
            "title": "Margin Mm"
          },
          "stock_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock File Id"
          },
          "parts_config": {
            "items": {
              "$ref": "#/components/schemas/RunConfigPartEntry"
            },
            "type": "array",
            "title": "Parts Config"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "created_by",
          "schema_version",
          "seed",
          "time_limit_s",
          "spacing_mm",
          "margin_mm",
          "parts_config"
        ],
        "title": "RunConfigResponse"
      },
      "RunCreateRequest": {
        "properties": {
          "run_config_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Config Id"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InlineRunConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunCreateRequest"
      },
      "RunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "RunListResponse"
      },
...
      "RunMetrics": {
        "properties": {
          "placements_count": {
            "type": "integer",
            "title": "Placements Count"
          },
          "unplaced_count": {
            "type": "integer",
            "title": "Unplaced Count"
          },
          "sheet_count": {
            "type": "integer",
            "title": "Sheet Count"
          }
        },
        "type": "object",
        "required": [
          "placements_count",
          "unplaced_count",
          "sheet_count"
        ],
        "title": "RunMetrics"
      },
      "RunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "run_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Config Id"
          },
          "triggered_by": {
            "type": "string",
            "title": "Triggered By"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "queued_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queued At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "solver_exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solver Exit Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunMetrics"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "triggered_by",
          "status"
        ],
        "title": "RunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}